train_model.py
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report
import joblib

df = pd.read_csv("synthetic_productivity.csv")

X = df.drop("productivity", axis=1)
y = df["productivity"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

model = RandomForestClassifier(n_estimators=200, random_state=42)
model.fit(X_train, y_train)

y_pred = model.predict(X_test)
print("Model trained successfully")
print(classification_report(y_test, y_pred))

joblib.dump(model, "model/model.pkl")
print("Model saved at model/model.pkl")
