requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
requests
joblib
python-dotenv
sqlalchemy
